--- src/main/java/com/example/jobgsm/domain/signup/exception/ExistEmailException.java
@@ -1,15 +1,0 @@
-package com.example.jobgsm.domain.signup.exception;
-
-import com.example.jobgsm.global.exception.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class ExistEmailException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public ExistEmailException(String message) {
-        super(message);
-        this.errorCode = ErrorCode.ALREADY_EXIST_EMAIL;
-    }
-}
