--- src/main/java/com/example/jobgsm/domain/signup/exception/MemberNotFoundException.java
@@ -1,16 +1,0 @@
-package com.example.jobgsm.domain.signup.exception;
-
-
-import com.example.jobgsm.global.exception.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class MemberNotFoundException extends RuntimeException{
-
-    private final ErrorCode errorCode;
-
-    public MemberNotFoundException(String message) {
-        super(message);
-        this.errorCode = ErrorCode.MEMBER_NOT_FOUND;
-    }
-}
