--- src/main/java/com/example/jobgsm/domain/signup/exception/RefreshTokenNotFoundException.java
@@ -1,15 +1,0 @@
-package com.example.jobgsm.domain.signup.exception;
-
-import com.example.jobgsm.global.exception.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class RefreshTokenNotFoundException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public RefreshTokenNotFoundException(String message) {
-        super(message);
-        this.errorCode = ErrorCode.REFRESH_TOKEN_NOT_FOUND;
-    }
-
-}
