--- src/main/java/com/example/jobgsm/domain/signup/presentation/dto/request/MemberSignInRequestDto.java
@@ -1,24 +1,0 @@
-package com.example.jobgsm.domain.signup.presentation.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MemberSignInRequestDto {
-
-
-    @NotBlank
-    private String email;
-
-    @NotBlank
-    private String password;
-
-}
