--- src/main/java/com/example/jobgsm/domain/signup/presentation/dto/response/MemberSignUpResponseDto.java
@@ -1,18 +1,0 @@
-package com.example.jobgsm.domain.signup.presentation.dto.response;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class MemberSignUpResponseDto {
-    private String email;
-    private String password;
-    private String name;
-    private String grade;
-}
