+++ src/main/java/com/example/jobgsm/domain3/signup/exception/MemberNotFoundException.java
@@ -1,0 +1,16 @@
+package com.example.jobgsm.domain3.signup.exception;
+
+
+import com.example.jobgsm.global3.exception.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class MemberNotFoundException extends RuntimeException{
+
+    private final ErrorCode errorCode;
+
+    public MemberNotFoundException(String message) {
+        super(message);
+        this.errorCode = ErrorCode.MEMBER_NOT_FOUND;
+    }
+}
