+++ src/main/java/com/example/jobgsm/domain3/signup/presentation/dto/request/MemberSignInRequestDto.java
@@ -1,0 +1,21 @@
+package com.example.jobgsm.domain3.signup.presentation.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class MemberSignInRequestDto {
+
+
+    @NotBlank
+    private String email;
+
+    @NotBlank
+    private String password;
+
+}
