+++ src/main/java/com/example/jobgsm/domain3/signup/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.example.jobgsm.domain3.signup.repository;
+
+import com.example.jobgsm.domain3.signup.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Integer> {
+
+    Optional<Member> findByEmail(String loginId);
+
+}
