--- src/main/java/com/example/jobgsm/global/exception/ErrorMessage.java
@@ -1,11 +1,0 @@
-package com.example.jobgsm.global.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class ErrorMessage {
-    private final String message;
-    private final int status;
-}
