--- src/main/java/com/example/jobgsm/global/role/Role.java
@@ -1,5 +1,0 @@
-package com.example.jobgsm.global.role;
-
-public enum Role {
-    USER, ADMIN
-}
