+++ src/main/java/com/example/jobgsm/global3/exception/ErrorMessage.java
@@ -1,0 +1,11 @@
+package com.example.jobgsm.global3.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ErrorMessage {
+    private final String message;
+    private final int status;
+}
