+++ src/main/java/com/example/jobgsm/global3/role/Role.java
@@ -1,0 +1,5 @@
+package com.example.jobgsm.global3.role;
+
+public enum Role {
+    USER, ADMIN
+}
