+++ src/main/java/com/example/jobgsm/global3/security/config/SecurityConfig.java
@@ -1,0 +1,38 @@
+package com.example.jobgsm.global3.security.config;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+
+@EnableWebSecurity
+@RequiredArgsConstructor
+@Configuration
+public class SecurityConfig {
+
+    @Bean
+    protected SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .authorizeRequests()
+                .antMatchers( "/member/login", "/member/singUp").permitAll()
+                .and()
+                .formLogin()
+                .loginPage("/member/login")
+                .loginProcessingUrl("/login_proc")
+                .defaultSuccessUrl("/member/login")
+                .failureUrl("/member/login")
+                .and()
+                .csrf().disable();
+
+        return http.build();
+
+    }
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
+    }
+}
